backups-fast-mcp-v1/pyproject.toml
[project]
name = "boilerplate-fast-mcp"
version = "0.1.0"
description = "Boilerplate app to create MCP servers with FastMCP"
authors = [{ name = "Marcelo Mastrangi Bracali" }]
readme = "README.md"
requires-python = ">=3.12, <3.15"
dependencies = [
    "pyfiglet (>=1.0.4,<2.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "wsproto (>=1.2.0)",
    "requests (>=2.32.0,<3.0.0)",
    "pyfinance (>=1.3.0,<2.0.0)",
    "yfinance (>=1.2.0,<2.0.0)",
    "fastmcp (>=3.1.0,<4.0.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "python-frontmatter (>=1.1.0,<2.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning:aiohttp"]
addopts = "--no-header --tb=short"
